backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pyodbc
python-dotenv
schedule
pytz
jinja2